Fix conversations table migration

+++ database/migrations/2022_08_16_095504_update_converstions_table.php
@@ -14,17 +14,39 @@ class UpdateConverstionsTable extends Migration
     public function up()
     {
         Schema::table('conversations', function (Blueprint $table) {
-            $table->renameColumn('user_id','sender_id');
-            $table->string('sender_type');
-            $table->foreignId('receiver_id');
-            $table->string('receiver_type');
-            $table->foreignId('last_message_id')->nullable();
-            $table->timestamp('last_message_time')->nullable();
-            $table->integer('unread_message_count')->default(0);
-            $table->dropColumn('message');
-            $table->dropColumn('reply');
-            $table->dropColumn('checked');
-            $table->dropColumn('image');
+            if (Schema::hasColumn('conversations', 'user_id')) {
+                $table->renameColumn('user_id','sender_id');
+            }
+            if (!Schema::hasColumn('conversations', 'sender_type')) {
+                $table->string('sender_type');
+            }
+            if (!Schema::hasColumn('conversations', 'receiver_id')) {
+                $table->foreignId('receiver_id');
+            }
+            if (!Schema::hasColumn('conversations', 'receiver_type')) {
+                $table->string('receiver_type');
+            }
+            if (!Schema::hasColumn('conversations', 'last_message_id')) {
+                $table->foreignId('last_message_id')->nullable();
+            }
+            if (!Schema::hasColumn('conversations', 'last_message_time')) {
+                $table->timestamp('last_message_time')->nullable();
+            }
+            if (!Schema::hasColumn('conversations', 'unread_message_count')) {
+                $table->integer('unread_message_count')->default(0);
+            }
+            if (Schema::hasColumn('conversations', 'message')) {
+                $table->dropColumn('message');
+            }
+            if (Schema::hasColumn('conversations', 'reply')) {
+                $table->dropColumn('reply');
+            }
+            if (Schema::hasColumn('conversations', 'checked')) {
+                $table->dropColumn('checked');
+            }
+            if (Schema::hasColumn('conversations', 'image')) {
+                $table->dropColumn('image');
+            }
         });
     }
